+++ google/cloud/bigtable/internal/bulk_mutator.cc
@@ -190,8 +190,10 @@ std::vector<bigtable::FailedMutation> BulkMutatorState::OnRetryDone() && {
 BulkMutator::BulkMutator(std::string const& app_profile_id,
                          std::string const& table_name,
                          bigtable::IdempotentMutationPolicy& idempotent_policy,
-                         bigtable::BulkMutation mut)
-    : state_(app_profile_id, table_name, idempotent_policy, std::move(mut)) {}
+                         bigtable::BulkMutation mut,
+                         std::shared_ptr<OperationContext> operation_context)
+    : state_(app_profile_id, table_name, idempotent_policy, std::move(mut)),
+      operation_context_(std::move(operation_context)) {}
 
 grpc::Status BulkMutator::MakeOneRequest(bigtable::DataClient& client,
                                          grpc::ClientContext& client_context) {
@@ -215,10 +217,10 @@ Status BulkMutator::MakeOneRequest(BigtableStub& stub,
   // Send the request to the server.
   auto const& mutations = state_.BeforeStart();
 
-  // Configure the context
-  auto context = std::make_shared<grpc::ClientContext>();
-  google::cloud::internal::ConfigureContext(*context, options);
-  retry_context_.PreCall(*context);
+  // Configure the client_context
+  auto client_context = std::make_shared<grpc::ClientContext>();
+  google::cloud::internal::ConfigureContext(*client_context, options);
+  operation_context_->PreCall(*client_context);
   bool enable_server_retries = options.get<EnableServerRetriesOption>();
 
   struct UnpackVariant {
@@ -241,11 +243,11 @@ Status BulkMutator::MakeOneRequest(BigtableStub& stub,
   limiter.Acquire();
 
   // Read the stream of responses.
-  auto stream = stub.MutateRows(context, options, mutations);
+  auto stream = stub.MutateRows(client_context, options, mutations);
   while (absl::visit(UnpackVariant{state_, limiter, enable_server_retries},
                      stream->Read())) {
   }
-  retry_context_.PostCall(*context);
+  operation_context_->PostCall(*client_context, state_.last_status());
   return state_.last_status();
 }