+++ google/cloud/config/v1/config_connection.h
@@ -311,6 +311,20 @@ class ConfigConnection {
   GetTerraformVersion(
       google::cloud::config::v1::GetTerraformVersionRequest const& request);
 
+  virtual StreamRange<google::cloud::config::v1::ResourceChange>
+  ListResourceChanges(
+      google::cloud::config::v1::ListResourceChangesRequest request);
+
+  virtual StatusOr<google::cloud::config::v1::ResourceChange> GetResourceChange(
+      google::cloud::config::v1::GetResourceChangeRequest const& request);
+
+  virtual StreamRange<google::cloud::config::v1::ResourceDrift>
+  ListResourceDrifts(
+      google::cloud::config::v1::ListResourceDriftsRequest request);
+
+  virtual StatusOr<google::cloud::config::v1::ResourceDrift> GetResourceDrift(
+      google::cloud::config::v1::GetResourceDriftRequest const& request);
+
   virtual StreamRange<google::cloud::location::Location> ListLocations(
       google::cloud::location::ListLocationsRequest request);