+++ google/cloud/storage/internal/connection_impl.h
@@ -18,6 +18,7 @@
 #include "google/cloud/storage/idempotency_policy.h"
 #include "google/cloud/storage/internal/generic_stub.h"
 #include "google/cloud/storage/internal/storage_connection.h"
+#include "google/cloud/storage/object_read_stream.h"
 #include "google/cloud/storage/retry_policy.h"
 #include "google/cloud/storage/version.h"
 #include "google/cloud/internal/invocation_id_generator.h"
@@ -99,6 +100,18 @@ class StorageConnectionImpl
       DeleteResumableUploadRequest const& request) override;
   StatusOr<QueryResumableUploadResponse> UploadChunk(
       UploadChunkRequest const& request) override;
+  StatusOr<std::unique_ptr<std::string>> UploadFileSimple(
+      std::string const& file_name, std::size_t file_size,
+      InsertObjectMediaRequest& request) override;
+  StatusOr<std::unique_ptr<std::istream>> UploadFileResumable(
+      std::string const& file_name, ResumableUploadRequest& request) override;
+  Status DownloadStreamToFile(ObjectReadStream&& stream,
+                              std::string const& file_name,
+                              ReadObjectRangeRequest const& request) override;
+  StatusOr<ObjectMetadata> ExecuteParallelUploadFile(
+      std::vector<std::thread> threads,
+      std::vector<ParallelUploadFileShard> shards,
+      bool ignore_cleanup_failures) override;
 
   StatusOr<ListBucketAclResponse> ListBucketAcl(
       ListBucketAclRequest const& request) override;