+++ google/cloud/video/transcoder/v1/transcoder_client.h
@@ -358,7 +358,7 @@ class TranscoderServiceClient {
   /// [`StatusOr`]: @ref google::cloud::StatusOr
   /// [`Status`]: @ref google::cloud::Status
   /// [google.cloud.video.transcoder.v1.CreateJobTemplateRequest]: @googleapis_reference_link{google/cloud/video/transcoder/v1/services.proto#L194}
-  /// [google.cloud.video.transcoder.v1.JobTemplate]: @googleapis_reference_link{google/cloud/video/transcoder/v1/resources.proto#L162}
+  /// [google.cloud.video.transcoder.v1.JobTemplate]: @googleapis_reference_link{google/cloud/video/transcoder/v1/resources.proto#L168}
   ///
   // clang-format on
   StatusOr<google::cloud::video::transcoder::v1::JobTemplate> CreateJobTemplate(
@@ -390,7 +390,7 @@ class TranscoderServiceClient {
   /// [`StatusOr`]: @ref google::cloud::StatusOr
   /// [`Status`]: @ref google::cloud::Status
   /// [google.cloud.video.transcoder.v1.CreateJobTemplateRequest]: @googleapis_reference_link{google/cloud/video/transcoder/v1/services.proto#L194}
-  /// [google.cloud.video.transcoder.v1.JobTemplate]: @googleapis_reference_link{google/cloud/video/transcoder/v1/resources.proto#L162}
+  /// [google.cloud.video.transcoder.v1.JobTemplate]: @googleapis_reference_link{google/cloud/video/transcoder/v1/resources.proto#L168}
   ///
   // clang-format on
   StatusOr<google::cloud::video::transcoder::v1::JobTemplate> CreateJobTemplate(
@@ -426,7 +426,7 @@ class TranscoderServiceClient {
   /// [`future`]: @ref google::cloud::future
   /// [`StatusOr`]: @ref google::cloud::StatusOr
   /// [`Status`]: @ref google::cloud::Status
-  /// [google.cloud.video.transcoder.v1.JobTemplate]: @googleapis_reference_link{google/cloud/video/transcoder/v1/resources.proto#L162}
+  /// [google.cloud.video.transcoder.v1.JobTemplate]: @googleapis_reference_link{google/cloud/video/transcoder/v1/resources.proto#L168}
   /// [google.cloud.video.transcoder.v1.ListJobTemplatesRequest]: @googleapis_reference_link{google/cloud/video/transcoder/v1/services.proto#L216}
   ///
   // clang-format on
@@ -465,7 +465,7 @@ class TranscoderServiceClient {
   /// [`future`]: @ref google::cloud::future
   /// [`StatusOr`]: @ref google::cloud::StatusOr
   /// [`Status`]: @ref google::cloud::Status
-  /// [google.cloud.video.transcoder.v1.JobTemplate]: @googleapis_reference_link{google/cloud/video/transcoder/v1/resources.proto#L162}
+  /// [google.cloud.video.transcoder.v1.JobTemplate]: @googleapis_reference_link{google/cloud/video/transcoder/v1/resources.proto#L168}
   /// [google.cloud.video.transcoder.v1.ListJobTemplatesRequest]: @googleapis_reference_link{google/cloud/video/transcoder/v1/services.proto#L216}
   ///
   // clang-format on
@@ -495,7 +495,7 @@ class TranscoderServiceClient {
   /// [`StatusOr`]: @ref google::cloud::StatusOr
   /// [`Status`]: @ref google::cloud::Status
   /// [google.cloud.video.transcoder.v1.GetJobTemplateRequest]: @googleapis_reference_link{google/cloud/video/transcoder/v1/services.proto#L243}
-  /// [google.cloud.video.transcoder.v1.JobTemplate]: @googleapis_reference_link{google/cloud/video/transcoder/v1/resources.proto#L162}
+  /// [google.cloud.video.transcoder.v1.JobTemplate]: @googleapis_reference_link{google/cloud/video/transcoder/v1/resources.proto#L168}
   ///
   // clang-format on
   StatusOr<google::cloud::video::transcoder::v1::JobTemplate> GetJobTemplate(
@@ -525,7 +525,7 @@ class TranscoderServiceClient {
   /// [`StatusOr`]: @ref google::cloud::StatusOr
   /// [`Status`]: @ref google::cloud::Status
   /// [google.cloud.video.transcoder.v1.GetJobTemplateRequest]: @googleapis_reference_link{google/cloud/video/transcoder/v1/services.proto#L243}
-  /// [google.cloud.video.transcoder.v1.JobTemplate]: @googleapis_reference_link{google/cloud/video/transcoder/v1/resources.proto#L162}
+  /// [google.cloud.video.transcoder.v1.JobTemplate]: @googleapis_reference_link{google/cloud/video/transcoder/v1/resources.proto#L168}
   ///
   // clang-format on
   StatusOr<google::cloud::video::transcoder::v1::JobTemplate> GetJobTemplate(